fix style camelCase
