--- src/Data.json
@@ -1,176 +1,0 @@
-[
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 5,
-    "nameProduct": "1. Test Product",
-    "price": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 5,
-    "nameProduct": "1. Test Product",
-    "price": 5,
-    "discountPrice": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 3,
-    "nameProduct": "1. Test Product",
-    "price": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 4,
-    "nameProduct": "1. Test Product",
-    "price": 8,
-    "discountPrice": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 2,
-    "nameProduct": "1. Test Product",
-    "price": 10,
-    "discountPrice": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 1,
-    "nameProduct": "1. Test Product",
-    "price": 200
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 2,
-    "nameProduct": "1. Test Product",
-    "price": 150
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 3,
-    "nameProduct": "1. Test Product",
-    "price": 9,
-    "discountPrice": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 4,
-    "nameProduct": "1. Test Product",
-    "price": 7,
-    "discountPrice": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 5,
-    "nameProduct": "1. Test Product",
-    "price": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 2,
-    "nameProduct": "1. Test Product",
-    "price": 12,
-    "discountPrice": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 3,
-    "nameProduct": "1. Test Product",
-    "price": 28
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 4,
-    "nameProduct": "1. Test Product",
-    "price": 17,
-    "discountPrice": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 5,
-    "nameProduct": "1. Test Product",
-    "price": 5,
-    "discountPrice": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 1,
-    "nameProduct": "1. Test Product",
-    "price": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 2,
-    "nameProduct": "1. Test Product",
-    "price": 12,
-    "discountPrice": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 2,
-    "nameProduct": "1. Test Product",
-    "price": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 1,
-    "nameProduct": "1. Test Product",
-    "price": 10,
-    "discountPrice": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 3,
-    "nameProduct": "1. Test Product",
-    "price": 15,
-    "discountPrice": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 2,
-    "nameProduct": "1. Test Product",
-    "price": 13,
-    "discountPrice": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 1,
-    "nameProduct": "1. Test Product",
-    "price": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 2,
-    "nameProduct": "1. Test Product",
-    "price": 10,
-    "discountPrice": 20
-  },
-  {
-    "name": "Vendor 01",
-    "img": "./img-book.png",
-    "rating": 1,
-    "nameProduct": "1. Test Product",
-    "price": 20
-  }
-]
